pyproject.toml
[tool.poetry]
name = "drivers_venv"
version = "0.1.0"
description = ""
authors = ["LaKyXa <lakyxamersh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
django-admin = "^2.0.1"
djangorestframework = "^3.12.4"
Markdown = "^3.3.6"
pip = "^21.3.1"
drf-yasg = "^1.20.0"
django-filter = "^21.1"
black = "^21.12b0"
flake8 = "^4.0.1"
isort = "^5.10.1"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
